# string: optimize eq